requirements.txt
Flask==2.1.1
PyPDF2==3.0.0
langchain==0.0.123
langchain_google_genai==0.1.0
FAISS==1.7.2
python-dotenv==1.0.0
google-generativeai==0.2.0
AIzaSyBVL2rQFhcTilSKPFyeokyxWcFKYx7pKFk